Distributed Python2 and Python3 Xena commits
This patch provides a mostly working XenaPythonLib version with backwards
compatible Python3. The commits have been segregated into two separate fixes.
The first commit has all the flake8 fixes whereas the second commit has been
achieved with the help of Python's "2to3" utility[0]. In this commit, the
Python2 code has ben converted to Python3 code.

The unit tests seem to be improperly functioning at the moment, but we even
had some issues with python2, so deferring that for another work item.

v1->v2:
* Use send-email rather than pull-request
* separated the commits by type (flake8 and python3 conversion)

[0]: https://docs.python.org/2/library/2to3.html

The following flake8 ids are fixed by this commit.  This is in
preparation for a follow up commit which will add support for running
the xenalib under python3.

  E101, E111, E127, E128, E201, E202, E203, E211, E222, E251, E302, E303
  E305, E501, E703, F401, F821, F841, W191, W601

Signed-off-by: Nandan Kulkarni <nakulkar@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>

There is need for a python2 and python3 compatible XenaPythonLib,
since many distributions have been switching to default with Python3.

The main difference between Python2 and Python3 is the basic types
that exist to deal with texts and bytes. Python3 introduces more
distinguishing string types, but that means the "str" usage for
"bytes-on-the-wire" type programming will run into errors doing
comparisons.

Unfortunately, Python3 doesn't provide a convenient way to interpret
all strings as a particular type, so the conversion looks a bit ugly.

Some of this conversion was provided by the "2to3" utility.